git pull command hook for tracking a remote branch
this is useful if you want to track the current state of your server,
and expect devs to ssh into the server and run gitpull to update
It also tracks the local changes (only modified files),
tags the current commit, and pushes all of it back to the default remote.

Best part is it only takes over git pull if there's an "autotracking" branch defined for the current branch.
To do this, run `git config branch.<current branch>.autoTracking <remote branch>`.
For example: git config branch.production-server1.autoTracking release