aliases: add whichlink as a supercharged "which"
I frequently need to investigate what actual implementation of something
we're using. I could do this by just running "which ..." which will give
me the program we're running.

However, Ubuntu/Debian's alternatives system often buries the actual
executable under multiple symlinks. For example, the default ld on 16.04
is behind no less than 3 symlinks, and it doesn't even use the
alternatives system.

```
/usr/bin/ld -> x86_64-linux-gnu-ld*
/usr/bin/x86_64-linux-gnu-ld -> ld.bfd*
/usr/bin/ld.bfd -> x86_64-linux-gnu-ld.bfd*
/usr/bin/x86_64-linux-gnu-ld.bfd*
```

This can be useful for finding if there's a more logical symlink that
can be overridden, or if there is a broken symlink somewhere.