suppress tr errors in is_remote check
I recently moved to a new dev env on mac, and this was causing lots of noise because of two issues:

1. Apple's `w` emits a `\xff\xbf` sequence at the end of every line
2. Apple's `tr` does not handle those bytes, emits an error and promptly dies

Ultimately, this is a bandaid until I can fix the issue, but seeing as how macs are typically development workstations
and not typically accessed remotely (at least for my workflows), this is a low priority fix. In any case, it would only
produce incorrect results if the current tty is not the first line of output, which from what I can tell isn't the case,
and even if it were we assume it's local anyways, so it's not a priority to fix at all.