starship config
starship.rs offers a cross-platform, cross-shell prompt that is quick and efficient. It supports most of the things that
I've added to my own prompt library over the years, but even more than I could ever hope to maintain by myself.

There are a few things that could probably be improved (for example, remote tmux/etc session detection), but nothing
that is too difficult to add myself.