nano: include all the syntaxes that come with nano
this is a bit of pain because nano is installed by homebrew under /usr/local, but on debian/ubuntu under /usr.
Because nano barfs if any of the directories doesn't exist, we can only reasonably assume that /usr/local/share and
/usr/share exist, but not which one of them includes the nano directory. The wildcard thankfully suppresses the error,
but may cause issues if multiple versions of nano are installed as e.g. nano-5 and nano-4.

On the systems I use and manage this is not the case, but it may be someday, so I'll keep my eyes out for a way to fix
the ugly hack